Add files via upload
This repository includes files for creating a user database management system. On opening the home page it gives new user an option to register and existing users the option to login. On successful login it directs to user information display page, where one with admin rights can even delete data. On unsuccessful login it gives an error alert message.